fixed a bug for int in path.join in json2exp

+++ mescnn/detection/qupath/tile.py
@@ -9,7 +9,7 @@ if __name__ == '__main__':
     parser = argparse.ArgumentParser(description='Extract tiles from WSI')
     parser.add_argument('-w', '--wsi', type=str, help='path/to/wsi', required=True)
     parser.add_argument('-e', '--export', type=str, help='path/to/export', required=True)
-    parser.add_argument('--desired-op', type=int, help='Desired Magnification for performing segmentation', default=10)
+    parser.add_argument('--desired-op', type=float, help='Desired Magnification for performing segmentation', default=10)
     parser.add_argument('--tile-size', type=int, nargs=2, help='Size of a tile (in pixel, at maximum magnification available)', default=(4096, 4096))
     parser.add_argument('--tile-stride', type=int, nargs=2, help='Size of the stride between tiles (in pixel, at maximum magnification available)', default=(2048, 2048))
